+++ modules/alloy/src/demo/ts/ephox/alloy/demo/DemoSink.ts
@@ -10,7 +10,7 @@ const make = (): AlloyComponent => GuiFactory.build(
   Container.sketch({
     containerBehaviours: Behaviour.derive([
       Positioning.config({
-        useFixed: Fun.always
+        usePositioningType: Fun.constant('fixed')
       })
     ])
   })