+++ modules/alloy/src/test/ts/browser/ui/dropdown/DropdownListTest.ts
@@ -27,7 +27,7 @@ UnitTest.asynctest('Dropdown List', (success, failure) => {
     Container.sketch({
       containerBehaviours: Behaviour.derive([
         Positioning.config({
-          useFixed: Fun.always
+          usePositioningType: Fun.constant('fixed')
         })
       ])
     })