+++ modules/alloy/src/test/ts/browser/ui/dropdown/DropdownRefetchTest.ts
@@ -31,7 +31,7 @@ UnitTest.asynctest('DropdownRefetchTest', (success, failure) => {
     Container.sketch({
       containerBehaviours: Behaviour.derive([
         Positioning.config({
-          useFixed: Fun.always
+          usePositioningType: Fun.constant('fixed')
         })
       ])
     })