+++ modules/alloy/src/test/ts/browser/ui/dropdown/SplitDropdownTest.ts
@@ -23,7 +23,7 @@ UnitTest.asynctest('SplitDropdown List', (success, failure) => {
     Container.sketch({
       containerBehaviours: Behaviour.derive([
         Positioning.config({
-          useFixed: Fun.always
+          usePositioningType: Fun.constant('fixed')
         })
       ])
     })