+++ modules/alloy/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Changed
-- Changed positioning behaviour position style from `relative` to `absolute` to avoid Safari autoscroll. #TINY-9148
+- Changed `useFixed` function to `usePositioningType` to allow usage of `absolute` positioning which fixes autoscrolling in Safari. #TINY-9148
 
 ## 12.0.0 - 2022-11-23